+++ database/migrations/2023_07_03_160542_create_ppbje_details_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->bigInteger('ppbje_id');
             $table->bigInteger('item_id');
             $table->bigInteger('supplier_id');
-            $table->double('quantity');
-            $table->double('price');
-            $table->double('price_total');
+            $table->decimal('quantity', 12, 1);
+            $table->decimal('price', 12, 2);
+            $table->decimal('price_total', 12, 2);
             $table->bigInteger('purchase_id')->nullable();
             $table->bigInteger('receiving_id')->nullable();
             $table->timestamps();
